Migrate to com.otilm packages and SNAPSHOT dependencies
Bump parent dependencies to 1.4.2-SNAPSHOT and interfaces to
2.18.1-SNAPSHOT, whose Java namespace moved com.czertainly -> com.otilm.

- pom: change dependencies and interfaces groupId to com.otilm
- rewrite all imports com.czertainly.api / com.czertainly.core -> com.otilm
- move own package com.czertainly.np.email -> com.otilm.np.email
- update logback logger name and @ComponentScan base package
