+++ pom.xml
@@ -5,9 +5,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.czertainly</groupId>
+        <groupId>com.otilm</groupId>
         <artifactId>dependencies</artifactId>
-        <version>1.4.0</version>
+        <version>1.4.2-SNAPSHOT</version>
     </parent>
 
     <artifactId>email-notification-provider</artifactId>
@@ -16,7 +16,7 @@
 
     <properties>
         <!-- Temporary override for CVE-2026-41293 and CVE-2026-43512.
-          Remove after the parent com.czertainly:dependencies upgrades to a Spring Boot release that ships Tomcat 10.1.55+
+          Remove after the parent com.otilm:dependencies upgrades to a Spring Boot release that ships Tomcat 10.1.55+
           (expected with Spring Boot 3.5.15). Then delete this override. -->
         <tomcat.version>10.1.55</tomcat.version>
         
@@ -29,9 +29,9 @@
     <dependencies>
         <!-- CZERTAINLY -->
         <dependency>
-            <groupId>com.czertainly</groupId>
+            <groupId>com.otilm</groupId>
             <artifactId>interfaces</artifactId>
-            <version>2.18.0</version>
+            <version>2.18.1-SNAPSHOT</version>
         </dependency>
 
         <!-- https://mvnrepository.com/artifact/net.sf.kxml/kxml2 -->
