--- src/main/java/com/czertainly/np/email/api/AttributesControllerImpl.java
@@ -1,36 +1,0 @@
-package com.czertainly.np.email.api;
-
-import com.czertainly.api.exception.ValidationException;
-import com.czertainly.api.interfaces.connector.AttributesController;
-import com.czertainly.api.model.client.attribute.RequestAttribute;
-import com.czertainly.api.model.common.attribute.common.BaseAttribute;
-import com.czertainly.np.email.service.AttributeService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/v1/notificationProvider/{kind}/attributes")
-public class AttributesControllerImpl implements AttributesController {
-
-    private AttributeService attributeService;
-
-    @Autowired
-    public void setAttributeService(AttributeService attributeService) {
-        this.attributeService = attributeService;
-    }
-
-    @Override
-    public List<BaseAttribute> listAttributeDefinitions(@PathVariable String kind) {
-        return attributeService.getAttributes(kind);
-    }
-
-    @Override
-    public void validateAttributes(@PathVariable String kind, @RequestBody List<RequestAttribute> attributes) throws ValidationException {
-        attributeService.validateAttributes(kind, attributes);
-    }
-}
