--- src/main/java/com/czertainly/np/email/config/WebAppConfig.java
@@ -1,9 +1,0 @@
-package com.czertainly.np.email.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebAppConfig implements WebMvcConfigurer {
-
-}
