--- src/main/java/com/czertainly/np/email/service/HealthService.java
@@ -1,8 +1,0 @@
-package com.czertainly.np.email.service;
-
-import com.czertainly.api.model.common.HealthDto;
-
-public interface HealthService {
-
-    HealthDto checkHealth();
-}
