+++ src/main/java/com/otilm/np/email/api/NotificationInstanceControllerImpl.java
@@ -1,0 +1,79 @@
+package com.otilm.np.email.api;
+
+import com.otilm.api.exception.AlreadyExistException;
+import com.otilm.api.exception.NotFoundException;
+import com.otilm.api.exception.ValidationException;
+import com.otilm.api.interfaces.connector.NotificationInstanceController;
+import com.otilm.api.model.common.attribute.common.DataAttribute;
+import com.otilm.api.model.connector.notification.NotificationProviderInstanceDto;
+import com.otilm.api.model.connector.notification.NotificationProviderInstanceRequestDto;
+import com.otilm.api.model.connector.notification.NotificationProviderNotifyRequestDto;
+import com.otilm.np.email.service.AttributeService;
+import com.otilm.np.email.service.NotificationInstanceService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+import java.util.UUID;
+
+@RestController
+public class NotificationInstanceControllerImpl implements NotificationInstanceController {
+
+    private NotificationInstanceService notificationInstanceService;
+    private AttributeService attributeService;
+
+    @Autowired
+    public void setNotificationInstanceService(NotificationInstanceService notificationInstanceService) {
+        this.notificationInstanceService = notificationInstanceService;
+    }
+
+    @Autowired
+    public void setAttributeService(AttributeService attributeService) {
+        this.attributeService = attributeService;
+    }
+
+    //////// NotificationInstanceController implementation ////////
+
+    @Override
+    public List<NotificationProviderInstanceDto> listNotificationInstances() {
+        return notificationInstanceService.listNotificationInstances();
+    }
+
+    @Override
+    public NotificationProviderInstanceDto getNotificationInstance(String uuid) throws NotFoundException {
+        return notificationInstanceService.getNotificationInstance(UUID.fromString(uuid));
+    }
+
+    @Override
+    public NotificationProviderInstanceDto createNotificationInstance(NotificationProviderInstanceRequestDto request) throws AlreadyExistException {
+        if (!attributeService.validateAttributes(
+                request.getKind(), request.getAttributes())) {
+            throw new ValidationException("Notification instance attributes validation failed.");
+        }
+        return notificationInstanceService.createNotificationInstance(request);
+    }
+
+    @Override
+    public NotificationProviderInstanceDto updateNotificationInstance(String uuid, NotificationProviderInstanceRequestDto request) throws NotFoundException {
+        if (!attributeService.validateAttributes(
+                request.getKind(), request.getAttributes())) {
+            throw new ValidationException("Notification instance attributes validation failed.");
+        }
+        return notificationInstanceService.updateNotificationInstance(UUID.fromString(uuid), request);
+    }
+
+    @Override
+    public void removeNotificationInstance(String uuid) throws NotFoundException {
+        notificationInstanceService.removeNotificationInstance(UUID.fromString(uuid));
+    }
+
+    @Override
+    public void sendNotification(String uuid, NotificationProviderNotifyRequestDto request) throws NotFoundException {
+        notificationInstanceService.sendNotification(UUID.fromString(uuid), request);
+    }
+
+    @Override
+    public List<DataAttribute> listMappingAttributes(String kind) {
+        return attributeService.listMappingAttributes(kind);
+    }
+}
