+++ src/main/java/com/otilm/np/email/config/ApplicationConfig.java
@@ -1,0 +1,15 @@
+package com.otilm.np.email.config;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+@ComponentScan(basePackages = "com.otilm.np.email")
+public class ApplicationConfig {
+    private static final Logger logger = LoggerFactory.getLogger(ApplicationConfig.class);
+
+}
