+++ src/main/java/com/otilm/np/email/config/WebAppConfig.java
@@ -1,0 +1,9 @@
+package com.otilm.np.email.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebAppConfig implements WebMvcConfigurer {
+
+}
