+++ src/main/java/com/otilm/np/email/exception/NotificationException.java
@@ -1,0 +1,17 @@
+package com.otilm.np.email.exception;
+
+public class NotificationException extends RuntimeException {
+
+    public NotificationException() {
+        super();
+    }
+
+    public NotificationException(String message) {
+        super(message);
+    }
+
+    public NotificationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
