+++ src/main/java/com/otilm/np/email/service/AttributeService.java
@@ -1,0 +1,17 @@
+package com.otilm.np.email.service;
+
+import com.otilm.api.model.client.attribute.RequestAttribute;
+import com.otilm.api.model.common.attribute.common.BaseAttribute;
+import com.otilm.api.model.common.attribute.common.DataAttribute;
+
+import java.util.List;
+
+public interface AttributeService {
+
+    List<BaseAttribute> getAttributes(String kind);
+
+    boolean validateAttributes(String kind, List<RequestAttribute> attributes);
+
+    List<DataAttribute> listMappingAttributes(String kind);
+
+}
