+++ src/main/java/com/otilm/np/email/service/HealthService.java
@@ -1,0 +1,8 @@
+package com.otilm.np.email.service;
+
+import com.otilm.api.model.common.HealthDto;
+
+public interface HealthService {
+
+    HealthDto checkHealth();
+}
