+++ src/main/resources/logback-spring.xml
@@ -8,5 +8,5 @@
         <appender-ref ref="CONSOLE" />
 <!--        <appender-ref ref="FILE" />-->
     </root>
-    <logger name="com.czertainly" level="INFO"/>
+    <logger name="com.otilm" level="INFO"/>
 </configuration>
